+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - [feat:support dynamic multi-discovery.](https://github.com/Tencent/spring-cloud-tencent/pull/1591)
 - [feat:support ipv6.](https://github.com/Tencent/spring-cloud-tencent/pull/1592)
 - [feat:support config all recover enabled.](https://github.com/Tencent/spring-cloud-tencent/pull/1593)
+- [feat:support stat reporting path aggregation.](https://github.com/Tencent/spring-cloud-tencent/pull/1602)